emg tests
